refactor: split workflow task store
