+++ server/workflowstore/transitions.go
@@ -1,0 +1,71 @@
+package workflowstore
+
+import (
+	"context"
+	"database/sql"
+	"fmt"
+
+	"builder/server/metadata/sqlitegen"
+	"builder/server/workflow"
+)
+
+func (s *Store) ListTransitions(ctx context.Context, taskID workflow.TaskID) ([]TransitionRecord, error) {
+	rows, err := s.queries.ListTaskTransitions(ctx, string(taskID))
+	if err != nil {
+		return nil, err
+	}
+	out := make([]TransitionRecord, 0, len(rows))
+	for _, row := range rows {
+		outputs := map[string]string{}
+		if err := unmarshalJSON(row.OutputValuesJson, &outputs); err != nil {
+			return nil, err
+		}
+		out = append(out, TransitionRecord{ID: workflow.TransitionID(row.ID), TaskID: workflow.TaskID(row.TaskID), TransitionID: row.TransitionID, State: row.State, Commentary: row.Commentary, OutputValues: outputs, CreatedAt: row.CreatedAtUnixMs})
+	}
+	return out, nil
+}
+
+func (s *Store) ListTransitionEdges(ctx context.Context, transitionID workflow.TransitionID) ([]TransitionEdgeRecord, error) {
+	rows, err := s.queries.ListTaskTransitionEdges(ctx, string(transitionID))
+	if err != nil {
+		return nil, err
+	}
+	out := make([]TransitionEdgeRecord, 0, len(rows))
+	for _, row := range rows {
+		out = append(out, TransitionEdgeRecord{
+			ID:                   row.ID,
+			TaskTransitionID:     workflow.TransitionID(row.TaskTransitionID),
+			WorkflowEdgeID:       workflow.EdgeID(row.WorkflowEdgeID.String),
+			EdgeKey:              row.EdgeKey,
+			TargetNodeID:         workflow.NodeID(row.TargetNodeID.String),
+			TargetPlacementID:    workflow.PlacementID(row.TargetPlacementID.String),
+			State:                row.State,
+			WorkflowRevisionSeen: row.WorkflowRevisionSeen,
+		})
+	}
+	return out, nil
+}
+
+func insertTransitionEdgeSnapshot(ctx context.Context, q *sqlitegen.Queries, transitionID string, revision int64, edge edgeContractSnapshot, targetPlacementID string, state string) error {
+	if err := q.InsertTaskTransitionEdge(ctx, sqlitegen.InsertTaskTransitionEdgeParams{
+		ID:                     prefixedID("transition-edge"),
+		TaskTransitionID:       transitionID,
+		WorkflowEdgeID:         sql.NullString{String: string(edge.ID), Valid: edge.ID != ""},
+		EdgeKey:                string(edge.Key),
+		WorkflowRevisionSeen:   revision,
+		TargetNodeID:           sql.NullString{String: string(edge.TargetNode.ID), Valid: edge.TargetNode.ID != ""},
+		TargetNodeKey:          string(edge.TargetNode.Key),
+		TargetNodeDisplayName:  edge.TargetNode.DisplayName,
+		TargetNodeKind:         string(edge.TargetNode.Kind),
+		TargetPlacementID:      sql.NullString{String: targetPlacementID, Valid: targetPlacementID != ""},
+		State:                  state,
+		ContextMode:            string(edge.ContextMode),
+		RequiresApproval:       boolToInt64(edge.RequiresApproval),
+		InputBindingsJson:      mustJSON(edge.InputBindings),
+		OutputRequirementsJson: mustJSON(edge.OutputRequirements),
+		MetadataJson:           "{}",
+	}); err != nil {
+		return fmt.Errorf("insert transition edge snapshot: %w", err)
+	}
+	return nil
+}
